Figuring-out Google Sheet Import and Basic Styling
Figured-out the basic method of requesting the Google Sheet via axios and using its response to feed the CSV-to-JSON module.

Also added Semantic UI CSS and some basic boilerplate to generate a nice looking table.

My intention is to not use Redux or anything in this project because it's very simple.